Can now initialize w/ username/pass; also support for user.getapikey API call
Thanks to Aaron Hamid for providing an initial implementation for this,
and getting me off my butt to think about it.

I decided to take this a different direction for a few reasons:

 - I think that the api_key vs. username/password is primarily an initialization-
   time concern.

 - I prefer (when possible) to do nothing in initialize() other than set instance
   variables and, if needed, raise exceptions -- doing other things is hard to
   do TDD, as it's hard to isolate the functionality otherwise.

 - This means that whatever has to happen later will be lazy-loading of
   some sort or another.  Making a policy of "always use api key for remote
   calls" makes it possible to lazy-load the API key when only username/password
   is provided at initialization.

 - This makes the user.getapikey API call look odd in User, but it turns out that
   we can use that call transparently in the lazy loader, and provide just a simple
   accessor wrapper in User that gets the return value.

The net result is that when we have the API key we don't call remote to look it up.
When we don't have it (just username/password) we will call the user.getapikey method
to look it up (exactly once).  If the first call is to User#getapikey then we're
basically doing what one would want anyway, otherwise, it happened at some point and
we're just going to return the cached value.