Add an explicit test for the User-Agent
The User-Agent is covered by necessity by other tests, because
RSpec 2 does not support fuzzy matching for `Hash` arguments
in mocks.  It is a good idea to also have a dedicated test for
the User-Agent functionality so that, if it becomes possible
to use fuzzy-matching in these tests, we still have coverage
for the User-Agent functionality.